move carinfo sensors to main instead of full
