+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorReceiver.kt
@@ -58,6 +58,7 @@ class SensorReceiver : SensorReceiverBase() {
             AudioSensorManager(),
             BatterySensorManager(),
             BluetoothSensorManager(),
+            CarSensorManager(),
             DisplaySensorManager(),
             DNDSensorManager(),
             DynamicColorSensorManager(),
